requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.5.0
fitbit==0.3.1
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.3.1
isort==4.3.21
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.text==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.1
oauthlib==3.1.0
paho-mqtt==1.5.0
pandas==1.0.1
pkg-resources==0.0.0
portend==2.6
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tempora==2.1.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
zc.lockfile==2.0
zipp==3.1.0
